requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.1.0
backcall==0.2.0
decorator==5.1.1
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
executing==1.2.0
ipython==8.7.0
jedi==0.18.2
matplotlib-inline==0.1.6
numpy==1.23.5
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pip==22.3.1
prompt-toolkit==3.0.33
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
python-dateutil==2.8.2
pytz==2022.6
setuptools==58.0.4
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.5.0
wcwidth==0.2.5
